+++ js/unix_formatting.js
@@ -486,12 +486,7 @@
             return [styles.join(''), color, background];
         }
         return function from_ansi(input, options) {
-            if (!input.match(/\x1b/)) {
-                if (options.position) {
-                    return [input, options.position];
-                }
-                return input;
-            }
+            options = options || {};
             input = $.terminal.unescape_brackets(input);
             var settings = $.extend({
                 unixFormattingEscapeBrackets: false,